requirements.txt
Flask==1.1.2
Jinja2==2.11.3
Pillow==8.1.0
numpy==1.20.0
matplotlib==3.3.4
gunicorn==20.0.4